# specs/001-guest-lock-pins/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Guest Lock PIN Manager API
  description: |
    REST API for the Guest Lock PIN Manager Home Assistant addon.
    Manages calendar subscriptions, lock configurations, and PIN lifecycle.
  version: 1.0.0
  contact:
    name: Guest Lock Manager
servers:
  - url: /api
    description: Home Assistant ingress-proxied API
...
paths:
...
  /static-pins:
    get:
      tags: [static-pins]
      summary: List all static PINs
      operationId: listStaticPins
      responses:
        '200':
          description: List of static PINs
          content:
            application/json:
              schema:
                type: object
                properties:
                  pins:
                    type: array
                    items:
                      $ref: '#/components/schemas/StaticPin'
    post:
      tags: [static-pins]
      summary: Create a static PIN
      operationId: createStaticPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaticPinCreate'
      responses:
        '201':
          description: Static PIN created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaticPin'
        '400':
          $ref: '#/components/responses/BadRequest'
  /static-pins/{id}:
    parameters:
      - $ref: '#/components/parameters/StaticPinId'
    get:
      tags: [static-pins]
      summary: Get static PIN details
      operationId: getStaticPin
      responses:
        '200':
          description: Static PIN details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaticPin'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [static-pins]
      summary: Update static PIN
      operationId: updateStaticPin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StaticPinUpdate'
      responses:
        '200':
          description: Static PIN updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StaticPin'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [static-pins]
      summary: Delete static PIN
      operationId: deleteStaticPin
      responses:
        '204':
          description: Static PIN deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # ============== SYSTEM ==============
...
  /settings:
    get:
      tags: [system]
      summary: Get addon settings
      operationId: getSettings
      responses:
        '200':
          description: Current settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
    put:
      tags: [system]
      summary: Update addon settings
      operationId: updateSettings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsUpdate'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
components:
  parameters:
...
    StaticPinId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "not_found"
        message:
          type: string
          example: "Calendar not found"
        details:
          type: object
          additionalProperties: true
...
    PinLockStatus:
      type: object
      properties:
        lock_id:
          type: string
          format: uuid
        lock_name:
          type: string
        slot_number:
          type: integer
        sync_status:
          type: string
          enum: [pending, synced, failed, removed]
    StaticPin:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        pin_code:
          type: string
        enabled:
          type: boolean
        always_active:
          type: boolean
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
        locks:
          type: array
          items:
            $ref: '#/components/schemas/PinLockStatus'
        created_at:
          type: string
          format: date-time
...
    StaticPinCreate:
      type: object
      required: [name, pin_code, lock_ids]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        pin_code:
          type: string
          pattern: '^\d{4,8}$'
        always_active:
          type: boolean
          default: false
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleInput'
        lock_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
    StaticPinUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        pin_code:
          type: string
          pattern: '^\d{4,8}$'
        enabled:
          type: boolean
        always_active:
          type: boolean
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/ScheduleInput'
        lock_ids:
          type: array
          items:
            type: string
            format: uuid
    Schedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        day_name:
          type: string
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
    ScheduleInput:
      type: object
      required: [day_of_week, start_time, end_time]
      properties:
        day_of_week:
          type: integer
          minimum: 0
          maximum: 6
        start_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        end_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
...
    Settings:
      type: object
      properties:
        default_sync_interval_min:
          type: integer
        min_pin_length:
          type: integer
        max_pin_length:
          type: integer
        checkin_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        checkout_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        timezone:
          type: string
        battery_efficient_mode:
          type: boolean
        batch_window_seconds:
          type: integer
    SettingsUpdate:
      type: object
      properties:
        default_sync_interval_min:
          type: integer
          minimum: 5
        min_pin_length:
          type: integer
          minimum: 4
          maximum: 8
        max_pin_length:
          type: integer
          minimum: 4
          maximum: 8
        checkin_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        checkout_time:
          type: string
          pattern: '^\d{2}:\d{2}$'
        battery_efficient_mode:
          type: boolean
        batch_window_seconds:
          type: integer
          minimum: 10
          maximum: 120